z2_plus: audio_platform_info: Disable ACDB for OUT_BT_SCO_WB
The entry for this calibration seems to be badly read or
corrupted: that makes it the kernel to send a Q6 CVP
parameter that is setting 15 RX channels and that is
obviously ... wrong.

Disable the ACDB calibration for the BT SCO WB device
at least until the database gets eventually fixed.

Signed-off-by: Dmitrii <bankersenator@gmail.com>